Remove old resources
